Correct a comparison of integers of different signs warning.

+++ src/utils/cnpy.hpp
@@ -103,7 +103,7 @@ namespace cnpy
                 assert(tmp_dims == ndims);
             }
 
-            for (int i = 1; i < ndims; i++) {
+            for (size_t i = 1; i < ndims; i++) {
                 if (shape[i] != tmp_shape[i]) {
                     std::cout <<"libnpy error: npy_save attempting to append misshaped data to " << fname
                               << "\n";
@@ -125,7 +125,7 @@ namespace cnpy
         }
 
         unsigned int nels = 1;
-        for (int i = 0; i < ndims; ++i) nels *= shape[i];
+        for (size_t i = 0; i < ndims; ++i) nels *= shape[i];
 
         fwrite(data, sizeof(T), nels, fp);
         fclose(fp);
